+++ src/model/interfaces/ComplexNumberUI.ts
@@ -1,0 +1,6 @@
+export interface ComplexNumberUI {
+    realValue: number;
+    imaginaryValue: number;
+    setRealValue: any;
+    setImaginaryValue: any;
+}