+++ src/model/interfaces/NumberInputProps.ts
@@ -1,4 +1,6 @@
 export interface NumberInputProps {
   label: string;
   inputID: string;
+  setValue: any;
+  value: any;
 }