+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0 - 2021-03-21
+### Added
+- Converts complex numbers from rectangular form to polar
+- Defines links style
+- Links site title to homepage
+
 ## 0.1.1 - 2021-03-20
 ### Added
 - Refactores so much